requirements.txt
python-telegram-bot==20.7
playwright==1.40.0
pandas==2.1.4
openpyxl==3.1.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0
aiohttp==3.9.1
